+++ geniusyield-server-lib/src/GeniusYield/Server/Dex/PartialOrder.hs
@@ -4,6 +4,7 @@ module GeniusYield.Server.Dex.PartialOrder (
   OrderInfo (..),
   poiToOrderInfo,
   PodServerException (..),
+  PodOrderNotFound (..),
 ) where
 
 import Data.Aeson (ToJSON (..))
@@ -15,13 +16,14 @@ import Data.Swagger.Internal.Schema qualified as Swagger
 import Deriving.Aeson
 import Fmt
 import GHC.TypeLits (AppendSymbol, Symbol)
-import GeniusYield.Api.Dex.PartialOrder (PartialOrderInfo (..), cancelMultiplePartialOrders', fillMultiplePartialOrders', getPartialOrdersInfos, getPartialOrdersInfos', getVersionsInOrders, orderByNft, partialOrderPrice', placePartialOrder', preferentiallySelectLatestPocd, preferentiallySelectLatestVersion, roundFunctionForPOCVersion)
-import GeniusYield.Api.Dex.PartialOrderConfig (RefPocd (..), SomeRefPocd (SomeRefPocd), fetchPartialOrderConfig, fetchPartialOrderConfigs)
+import GeniusYield.Api.Dex.PartialOrder (PartialOrderInfo (..), cancelMultiplePartialOrders', fillMultiplePartialOrders', fillPartialOrder', getPartialOrdersInfos, getPartialOrdersInfos', getVersionsInOrders, orderByNft, partialOrderPrice', placePartialOrder'', preferentiallySelectLatestPocd, preferentiallySelectLatestVersion, roundFunctionForPOCVersion)
+import GeniusYield.Api.Dex.PartialOrderConfig (RefPocd (..), SomeRefPocd (SomeRefPocd), fetchPartialOrderConfig, fetchPartialOrderConfigs, selectRefPocd)
 import GeniusYield.HTTP.Errors
 import GeniusYield.OrderBot.Domain.Markets (OrderAssetPair (..))
 import GeniusYield.Scripts.Dex.PartialOrderConfig (PartialOrderConfigInfoF (..))
 import GeniusYield.Scripts.Dex.Version (POCVersion (POCVersion1_1))
 import GeniusYield.Server.Ctx
+import GeniusYield.Server.Orphans ()
 import GeniusYield.Server.Tx (handleTxSign, handleTxSubmit, throwNoSigningKeyError)
 import GeniusYield.Server.Utils (addSwaggerDescription, addSwaggerExample, dropSymbolAndCamelToSnake, logDebug, logInfo)
 import GeniusYield.Types
@@ -32,6 +34,12 @@ import RIO.NonEmpty qualified as NonEmpty
 import RIO.Text qualified as T
 import Servant
 
+{- $setup
+
+>>> :set -XOverloadedStrings -XTypeApplications
+>>> import qualified Data.Aeson                 as Aeson
+-}
+
 -- | Number of orders that we at most allow to be filled in a single transaction.
 maxFillOrders ∷ GYNatural
 maxFillOrders = 5
@@ -39,10 +47,24 @@ maxFillOrders = 5
 data PodServerException
   = -- | Cannot fill more than allowed number of orders.
     PodMultiFillMoreThanAllowed
-  | -- | When filling multiple orders, they all should have same payment token so that taker fee is charged in only one token.
-    PodMultiFillNotAllSamePaymentToken
   deriving stock (Show)
   deriving anyclass (Exception)
+
+-- | When order whose details is queried for is not found.
+data PodOrderNotFound = PodOrderNotFound
+  deriving (Eq, Show, Generic)
+  deriving anyclass (Exception, ToJSON)
+
+instance Swagger.ToSchema PodOrderNotFound where
+  declareNamedSchema =
+    Swagger.genericDeclareNamedSchema Swagger.defaultSchemaOptions
+      & addSwaggerDescription (toErrDescription PodOrderNotFound)
+
+class ErrDescription e where
+  toErrDescription ∷ e → Text
+
+instance ErrDescription PodOrderNotFound where
+  toErrDescription _ = "Order not found"
 
 instance IsGYApiError PodServerException where
   toApiError PodMultiFillMoreThanAllowed =
@@ -51,11 +73,13 @@ instance IsGYApiError PodServerException where
         gaeHttpStatus = status400,
         gaeMsg = T.pack $ "Orders to fill is more than " <> show maxFillOrders
       }
-  toApiError PodMultiFillNotAllSamePaymentToken =
+
+instance IsGYApiError PodOrderNotFound where
+  toApiError PodOrderNotFound =
     GYApiError
-      { gaeErrorCode = "MULTI_FILL_NOT_SAME_PAIR",
-        gaeHttpStatus = status400,
-        gaeMsg = "Given orders are not having same payment token"
+      { gaeErrorCode = "ORDER_NOT_FOUND",
+        gaeHttpStatus = status404,
+        gaeMsg = toErrDescription PodOrderNotFound
       }
 
 type OrderInfoPrefix ∷ Symbol
@@ -63,13 +87,16 @@ type OrderInfoPrefix = "oi"
 
 data OrderInfo = OrderInfo
   { oiOfferAmount ∷ !GYRational,
+    oiOfferAmountInDatum ∷ !GYNatural,
     oiPrice ∷ !GYRational,
+    oiPriceInDatum ∷ !Rational,
     oiStart ∷ !(Maybe GYTime),
     oiEnd ∷ !(Maybe GYTime),
     oiOwnerAddress ∷ !GYAddressBech32,
     oiOwnerKeyHash ∷ !GYPubKeyHash,
     oiOutputReference ∷ !GYTxOutRef,
-    oiNFTToken ∷ !GYAssetClass
+    oiNFTToken ∷ !GYAssetClass,
+    oiVersion ∷ !POCVersion
   }
   deriving stock (Generic)
   deriving
@@ -79,6 +106,7 @@ data OrderInfo = OrderInfo
 instance Swagger.ToSchema OrderInfo where
   declareNamedSchema =
     Swagger.genericDeclareNamedSchema Swagger.defaultSchemaOptions {Swagger.fieldLabelModifier = dropSymbolAndCamelToSnake @OrderInfoPrefix}
+      & addSwaggerDescription "Order details given as part of order-book (bid and ask). Note that \"price_in_datum\" is the price given in datum whereas \"price\" is the rounded price in terms of currency asset per commodity asset. I.e., for a sell order, \"price\" is rounded value of \"price_in_datum\" whereas for buy order, \"price\" is rounded value of reciprocal of \"price_in_datum\". Likewise, \"offer_amount_in_datum\" is the amount given in datum whereas \"offer_amount\" is the rounded value in terms of commodity asset."
 
 type OrderInfoDetailedPrefix ∷ Symbol
 type OrderInfoDetailedPrefix = "oid"
@@ -113,13 +141,16 @@ poiToOrderInfo PartialOrderInfo {..} oap =
       poiOfferedAmount' = fromIntegral poiOfferedAmount
    in OrderInfo
         { oiOfferAmount = if isSell then poiOfferedAmount' else poiOfferedAmount' * poiPrice,
+          oiOfferAmountInDatum = naturalFromGHC poiOfferedAmount,
           oiPrice = if isSell then poiPrice else 1 / poiPrice,
+          oiPriceInDatum = rationalToGHC poiPrice,
           oiStart = poiStart,
           oiEnd = poiEnd,
           oiOwnerAddress = addressToBech32 poiOwnerAddr,
           oiOwnerKeyHash = poiOwnerKey,
           oiOutputReference = poiRef,
-          oiNFTToken = GYToken poiNFTCS poiNFT
+          oiNFTToken = GYToken poiNFTCS poiNFT,
+          oiVersion = poiVersion
         }
         :!: isSell
 
@@ -169,11 +200,11 @@ newtype ChangeAddress = ChangeAddress GYAddressBech32
 instance Swagger.ToSchema ChangeAddress where
   declareNamedSchema _ = do
     addrBech32Schema ← Swagger.declareSchema (Proxy @GYAddressBech32)
-    return $
-      Swagger.named "ChangeAddress" $
-        addrBech32Schema
-          & Swagger.description
-            %~ (\mt → mt <> Just " This is used as a change address by our balancer to send left over funds from selected inputs. If not provided, first address from given addresses list is used instead.")
+    return
+      $ Swagger.named "ChangeAddress"
+      $ addrBech32Schema
+      & Swagger.description
+      %~ (\mt → mt <> Just " This is used as a change address by our balancer to send left over funds from selected inputs. If not provided, first address from given addresses list is used instead.")
 
 type PlaceOrderReqPrefix ∷ Symbol
 type PlaceOrderReqPrefix = "pop"
@@ -211,7 +242,8 @@ data PlaceOrderTransactionDetails = PlaceOrderTransactionDetails
     potdMakerOfferedPercentFee ∷ !GYRational,
     potdMakerOfferedPercentFeeAmount ∷ !GYNatural,
     potdLovelaceDeposit ∷ !GYNatural,
-    potdOrderRef ∷ !GYTxOutRef
+    potdOrderRef ∷ !GYTxOutRef,
+    potdNFTToken ∷ !GYAssetClass
   }
   deriving stock (Generic)
   deriving
@@ -294,6 +326,25 @@ instance Swagger.ToSchema FillOrderParameters where
       & addSwaggerDescription "Fill order(s) request parameters."
       & addSwaggerExample (toJSON $ FillOrderParameters {fopAddresses = pure "addr_test1qrsuhwqdhz0zjgnf46unas27h93amfghddnff8lpc2n28rgmjv8f77ka0zshfgssqr5cnl64zdnde5f8q2xt923e7ctqu49mg5", fopChangeAddress = Just (ChangeAddress "addr_test1qrsuhwqdhz0zjgnf46unas27h93amfghddnff8lpc2n28rgmjv8f77ka0zshfgssqr5cnl64zdnde5f8q2xt923e7ctqu49mg5"), fopCollateral = Just "4293386fef391299c9886dc0ef3e8676cbdbc2c9f2773507f1f838e00043a189#1", fopOrderReferencesWithAmount = ("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e1ea747#0", 100) :| [("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e144444#0", 100)]})
 
+-- >>> Aeson.encode . BotFillOrderParameters $ pure ("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e1ea747#0", 100)
+-- "{\"order_references_with_amount\":[[\"0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e1ea747#0\",\"100\"]]}"
+newtype BotFillOrderParameters = BotFillOrderParameters
+  { bfopOrderReferencesWithAmount ∷ NonEmpty (GYTxOutRef, GYNatural)
+  }
+  deriving stock (Show, Generic)
+  deriving
+    (FromJSON, ToJSON)
+    via CustomJSON '[FieldLabelModifier '[StripPrefix BotFillOrderReqPrefix, CamelToSnake]] BotFillOrderParameters
+
+type BotFillOrderReqPrefix ∷ Symbol
+type BotFillOrderReqPrefix = "bfop"
+
+instance Swagger.ToSchema BotFillOrderParameters where
+  declareNamedSchema =
+    Swagger.genericDeclareNamedSchema Swagger.defaultSchemaOptions {Swagger.fieldLabelModifier = dropSymbolAndCamelToSnake @BotFillOrderReqPrefix}
+      & addSwaggerDescription "Fill order(s) request parameters specialized towards configured bot."
+      & addSwaggerExample (toJSON $ BotFillOrderParameters {bfopOrderReferencesWithAmount = ("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e1ea747#0", 100) :| [("0018dbaa1611531b9f11a31765e8abe875f9c43750b82b5f321350f31e144444#0", 100)]})
+
 type FillOrderResPrefix ∷ Symbol
 type FillOrderResPrefix = "fotd"
 
@@ -303,7 +354,7 @@ data FillOrderTransactionDetails = FillOrderTransactionDetails
     fotdTransactionFee ∷ !GYNatural,
     fotdTakerLovelaceFlatFee ∷ !GYNatural,
     fotdTakerOfferedPercentFee ∷ !GYRational,
-    fotdTakerOfferedPercentFeeAmount ∷ !GYNatural
+    fotdTakerOfferedPercentFeeAmount ∷ !GYValue
   }
   deriving stock (Generic)
   deriving
@@ -317,6 +368,9 @@ instance Swagger.ToSchema FillOrderTransactionDetails where
 type CommonCollateralText ∷ Symbol
 type CommonCollateralText = "Note that if \"collateral\" field is not provided, then framework would try to pick collateral UTxO on it's own and in that case would also be free to spend it (i.e., would be made available to coin balancer)."
 
+type CommonSignText ∷ Symbol
+type CommonSignText = "This endpoint would also sign & submit the built transaction. It uses the signing key from configuration to compute for wallet address. If collateral is specified in the configuration, then it would be used for."
+
 type OrdersAPI =
   Summary "Build transaction to create order"
     :> Description ("Build a transaction to create an order. In case \"stakeAddress\" field is provided then order is placed at a mangled address having the given staking credential. " `AppendSymbol` CommonCollateralText)
@@ -325,7 +379,7 @@ type OrdersAPI =
     :> ReqBody '[JSON] PlaceOrderParameters
     :> Post '[JSON] PlaceOrderTransactionDetails
     :<|> Summary "Create an order"
-      :> Description "Create an order. This endpoint would also sign & submit the built transaction."
+      :> Description ("Create an order. " `AppendSymbol` CommonSignText `AppendSymbol` " \"stakeAddress\" field from configuration, if provided, is used to place order at a mangled address.")
       :> ReqBody '[JSON] BotPlaceOrderParameters
       :> Post '[JSON] PlaceOrderTransactionDetails
     :<|> Summary "Build transaction to cancel order(s)"
@@ -335,7 +389,7 @@ type OrdersAPI =
       :> ReqBody '[JSON] CancelOrderParameters
       :> Post '[JSON] CancelOrderTransactionDetails
     :<|> Summary "Cancel order(s)"
-      :> Description "Cancel order(s). This endpoint would also sign & submit the built transaction."
+      :> Description ("Cancel order(s). " `AppendSymbol` CommonSignText)
       :> ReqBody '[JSON] BotCancelOrderParameters
       :> Delete '[JSON] CancelOrderTransactionDetails
     :<|> Summary "Get order(s) details"
@@ -343,11 +397,21 @@ type OrdersAPI =
       :> "details"
       :> ReqBody '[JSON] [GYAssetClass]
       :> Post '[JSON] [OrderInfoDetailed]
+    :<|> Summary "Get order details"
+      :> Description "Get details of an order using it's unique NFT token. Note that each order is identified uniquely by an associated NFT token which can then later be used to retrieve it's details across partial fills."
+      :> "details"
+      :> Capture "nft-token" GYAssetClass
+      :> UVerb 'GET '[JSON] '[WithStatus 200 OrderInfoDetailed, WithStatus 404 PodOrderNotFound]
     :<|> Summary "Build transaction to fill order(s)"
       :> Description ("Build a transaction to fill order(s). " `AppendSymbol` CommonCollateralText)
       :> "tx"
       :> "build-fill"
       :> ReqBody '[JSON] FillOrderParameters
+      :> Post '[JSON] FillOrderTransactionDetails
+    :<|> Summary "Build transaction to fill order(s)"
+      :> Description ("Build a transaction to fill order(s). " `AppendSymbol` CommonSignText)
+      :> "fill"
+      :> ReqBody '[JSON] BotFillOrderParameters
       :> Post '[JSON] FillOrderTransactionDetails
 
 handleOrdersApi ∷ Ctx → ServerT OrdersAPI IO
@@ -357,7 +421,9 @@ handleOrdersApi ctx =
     :<|> handleCancelOrders ctx
     :<|> handleCancelOrdersAndSignSubmit ctx
     :<|> handleOrdersDetails ctx
+    :<|> handleOrderDetails ctx
     :<|> handleFillOrders ctx
+    :<|> handleFillOrdersAndSignSubmit ctx
 
 handlePlaceOrder ∷ Ctx → PlaceOrderParameters → IO PlaceOrderTransactionDetails
 handlePlaceOrder ctx@Ctx {..} pops@PlaceOrderParameters {..} = do
@@ -368,11 +434,12 @@ handlePlaceOrder ctx@Ctx {..} pops@PlaceOrderParameters {..} = do
       pocVersion = POCVersion1_1
   SomeRefPocd (RefPocd (cfgRef :!: pocd)) ← runQuery ctx $ fetchPartialOrderConfig pocVersion porefs
   let unitPrice =
-        rationalFromGHC $
-          toInteger popPriceAmount % toInteger popOfferAmount
-  txBody ←
-    runSkeletonI ctx (NonEmpty.toList popAddresses') changeAddr popCollateral $
-      placePartialOrder'
+        rationalFromGHC
+          $ toInteger popPriceAmount
+          % toInteger popOfferAmount
+  (nftAC, txBody) ←
+    runSkeletonF ctx (NonEmpty.toList popAddresses') changeAddr popCollateral
+      $ placePartialOrder''
         porefs
         changeAddr
         (naturalToGHC popOfferAmount, popOfferToken)
@@ -395,7 +462,8 @@ handlePlaceOrder ctx@Ctx {..} pops@PlaceOrderParameters {..} = do
         potdMakerOfferedPercentFee = 100 * pociMakerFeeRatio pocd,
         potdMakerOfferedPercentFeeAmount = roundFunctionForPOCVersion pocVersion $ toRational popOfferAmount * rationalToGHC (pociMakerFeeRatio pocd),
         potdLovelaceDeposit = fromIntegral $ pociMinDeposit pocd,
-        potdOrderRef = txOutRefFromTuple (txId, 0)
+        potdOrderRef = txOutRefFromTuple (txId, 0),
+        potdNFTToken = nftAC
       }
 
 resolveCtxSigningKeyInfo ∷ Ctx → IO (Strict.Pair GYSomePaymentSigningKey GYAddress)
@@ -440,14 +508,24 @@ handleCancelOrdersAndSignSubmit ctx BotCancelOrderParameters {..} = do
   -- Though transaction id would be same, but we are returning it again, just in case...
   pure $ details {cotdTransactionId = txId, cotdTransaction = signedTx}
 
+handleOrderDetails ∷ Ctx → GYAssetClass → IO (Union '[WithStatus 200 OrderInfoDetailed, WithStatus 404 PodOrderNotFound])
+handleOrderDetails ctx@Ctx {..} ac = do
+  logInfo ctx $ "Getting order details for NFT token: " +|| ac ||+ ""
+  let porefs = dexPORefs ctxDexInfo
+  os ← runQuery ctx $ fmap poiToOrderInfoDetailed <$> orderByNft porefs ac
+  case os of
+    Nothing → throwIO PodOrderNotFound -- We could use `respond` here as well but then as it would not have @application/json@ header, it would not be caught by our @errorJsonWrapMiddleware@.
+    Just o → respond (WithStatus @200 o)
+
 handleOrdersDetails ∷ Ctx → [GYAssetClass] → IO [OrderInfoDetailed]
 handleOrdersDetails ctx@Ctx {..} acs = do
   logInfo ctx $ "Getting orders details for NFT tokens: " +|| acs ||+ ""
   let porefs = dexPORefs ctxDexInfo
   os ← forM acs $ \ac → do
     logDebug ctx $ "Getting order details for NFT token: " +|| ac ||+ ""
-    runQuery ctx $
-      fmap poiToOrderInfoDetailed <$> orderByNft porefs ac
+    runQuery ctx
+      $ fmap poiToOrderInfoDetailed
+      <$> orderByNft porefs ac
   pure $ catMaybes os
 
 handleFillOrders ∷ Ctx → FillOrderParameters → IO FillOrderTransactionDetails
@@ -465,11 +543,10 @@ handleFillOrders ctx@Ctx {..} fops@FillOrderParameters {..} = do
       maxFlatTakerFee = foldl' (\prevMax (poi, _) → max prevMax $ poiTakerLovelaceFlatFee poi) 0 ordersWithTokenBuyAmount
       fopAddresses' = addressFromBech32 <$> fopAddresses
       changeAddr = maybe (NonEmpty.head fopAddresses') (\(ChangeAddress addr) → addressFromBech32 addr) fopChangeAddress
-  takerFee' ← case valueToList takerFee of
-    [(_, feeAmt)] → pure $ fromIntegral feeAmt
-    _ → throwIO PodMultiFillNotAllSamePaymentToken
   txBody ← runSkeletonI ctx (NonEmpty.toList fopAddresses') changeAddr fopCollateral $ do
-    fillMultiplePartialOrders' porefs ordersWithTokenBuyAmount (Just refPocds) takerFee
+    case ordersWithTokenBuyAmount of
+      [(oi, amt)] → fillPartialOrder' porefs oi amt (Just $ selectRefPocd refPocds overallPocVersion) (fromIntegral $ valueAssetClass takerFee (poiAskedAsset oi))
+      _ → fillMultiplePartialOrders' porefs ordersWithTokenBuyAmount (Just refPocds) takerFee
   pure
     FillOrderTransactionDetails
       { fotdTransaction = unsignedTx txBody,
@@ -477,7 +554,7 @@ handleFillOrders ctx@Ctx {..} fops@FillOrderParameters {..} = do
         fotdTransactionFee = fromIntegral $ txBodyFee txBody,
         fotdTakerLovelaceFlatFee = fromIntegral maxFlatTakerFee,
         fotdTakerOfferedPercentFee = 100 * takerFeeRatio,
-        fotdTakerOfferedPercentFeeAmount = takerFee'
+        fotdTakerOfferedPercentFeeAmount = takerFee
       }
  where
   computePercentTakerFees ∷ Foldable t ⇒ POCVersion → t (PartialOrderInfo, Natural) → GYRational → GYValue
@@ -493,3 +570,13 @@ handleFillOrders ctx@Ctx {..} fops@FillOrderParameters {..} = do
         takerFee =
           Map.foldlWithKey' (\acc ac amt → acc <> valueSingleton ac (roundFunctionForPOCVersion overallPocVersion $ toRational amt * rationalToGHC takerFeeRatio)) mempty takerACWithAmt
      in takerFee
+
+handleFillOrdersAndSignSubmit ∷ Ctx → BotFillOrderParameters → IO FillOrderTransactionDetails
+handleFillOrdersAndSignSubmit ctx BotFillOrderParameters {..} = do
+  logInfo ctx "Filling order(s) and signing & submitting the transaction."
+  ctxAddr ← addressToBech32 <$> resolveCtxAddr ctx
+  details ← handleFillOrders ctx $ FillOrderParameters {fopAddresses = pure ctxAddr, fopChangeAddress = Just (ChangeAddress ctxAddr), fopCollateral = ctxCollateral ctx, fopOrderReferencesWithAmount = bfopOrderReferencesWithAmount}
+  signedTx ← handleTxSign ctx $ fotdTransaction details
+  txId ← handleTxSubmit ctx signedTx
+  -- Though transaction id would be same, but we are returning it again, just in case...
+  pure $ details {fotdTransactionId = txId, fotdTransaction = signedTx}