+++ geniusyield-server-lib/src/GeniusYield/Server/Run.hs
@@ -18,6 +18,11 @@ import GeniusYield.Server.Auth
 import GeniusYield.Server.Config (ServerConfig (..), coreConfigFromServerConfig, optionalSigningKeyFromServerConfig, serverConfigOptionalFPIO)
 import GeniusYield.Server.Constants (gitHash)
 import GeniusYield.Server.Ctx
+-- import RIO.ByteString.Lazy qualified as BL
+
+-- import Servant.PY (requests, writePythonForAPI)
+
+import GeniusYield.Server.Dex.HistoricalPrices.TapTools.Client (tapToolsClientEnv)
 import GeniusYield.Server.ErrorMiddleware
 import GeniusYield.Server.RequestLoggerMiddleware (gcpReqLogger)
 import GeniusYield.Server.Utils
@@ -27,10 +32,8 @@ import Network.Wai.Handler.Warp qualified as Warp
 import PackageInfo_geniusyield_server_lib qualified as PackageInfo
 import RIO hiding (Handler, logDebug, logErrorS, logInfo, logInfoS, onException)
 import RIO.ByteString qualified as B
--- import RIO.ByteString.Lazy qualified as BL
 import RIO.Text.Lazy qualified as LT
 import Servant
--- import Servant.PY (requests, writePythonForAPI)
 import Servant.Server.Experimental.Auth (AuthHandler)
 import Servant.Server.Internal.ServerError (responseServerError)
 import System.TimeManager (TimeoutThread (..))
@@ -39,6 +42,12 @@ runServer ∷ Maybe FilePath → IO ()
 runServer mfp = do
   serverConfig ← serverConfigOptionalFPIO mfp
   menv ← networkIdToMaestroEnv (case scMaestroToken serverConfig of Confidential t → t) (scNetworkId serverConfig)
+  mtenv ←
+    case scTapToolsApiKey serverConfig of
+      Nothing → pure Nothing
+      Just (Confidential apiKey) → do
+        tce ← tapToolsClientEnv
+        pure $ Just $ TapToolsEnv {tteClientEnv = tce, tteApiKey = apiKey}
   optionalSigningKey ← optionalSigningKeyFromServerConfig serverConfig
   let nid = scNetworkId serverConfig
       coreCfg = coreConfigFromServerConfig serverConfig
@@ -48,7 +57,7 @@ runServer mfp = do
         logErrorS = gyLogError providers mempty
     logInfoS $ "GeniusYield server version: " +| showVersion PackageInfo.version |+ "\nCommit used: " +| gitHash |+ "\nOptional collateral configuration: " +|| scCollateral serverConfig ||+ "\nAddress of optional wallet: " +|| fmap Strict.snd optionalSigningKey ||+ "\nOptional stake address: " +|| scStakeAddress serverConfig ||+ ""
     -- BL.writeFile "web/swagger/api.json" (encodePretty geniusYieldAPISwagger)
-    B.writeFile "web/swagger/api.yaml" (Yaml.encodePretty Yaml.defConfig geniusYieldAPISwagger)
+    B.writeFile "web/openapi/api.yaml" (Yaml.encodePretty Yaml.defConfig geniusYieldAPIOpenApi)
     reqLoggerMiddleware ← gcpReqLogger
     let
       -- These are only meant to catch fatal exceptions, application thrown exceptions should be caught beforehand.
@@ -78,22 +87,30 @@ runServer mfp = do
             ctxNetworkId = nid,
             ctxDexInfo =
               if
-                  | nid == GYMainnet → dexInfoDefaultMainnet
-                  | nid == GYTestnetPreprod → dexInfoDefaultPreprod
-                  | otherwise → error "Only mainnet & preprod network are supported",
+                | nid == GYMainnet → dexInfoDefaultMainnet
+                | nid == GYTestnetPreprod → dexInfoDefaultPreprod
+                | otherwise → error "Only mainnet & preprod network are supported",
             ctxMaestroProvider = MaestroProvider menv,
+            ctxTapToolsProvider = mtenv,
             ctxSigningKey = optionalSigningKey,
             ctxCollateral = scCollateral serverConfig,
             ctxStakeAddress = scStakeAddress serverConfig
           }
 
-    logInfoS $
-      "Starting GeniusYield server on port " +| scPort serverConfig |+ "\nCore config:\n" +| indentF 4 (fromString $ show coreCfg) |+ ""
-    Warp.runSettings settings . reqLoggerMiddleware . errLoggerMiddleware . errorJsonWrapMiddleware $
-      let context = apiKeyAuthHandler (case scServerApiKey serverConfig of Confidential t → apiKeyFromText t) :. EmptyContext
-       in serveWithContext mainAPI context
-            $ hoistServerWithContext
-              mainAPI
-              (Proxy ∷ Proxy '[AuthHandler Wai.Request ()])
-              (\ioAct → Handler . ExceptT $ first (apiErrorToServerError . exceptionHandler) <$> try ioAct)
-            $ mainServer ctx
+    logInfoS
+      $ "Starting GeniusYield server on port "
+      +| scPort serverConfig
+      |+ "\nCore config:\n"
+      +| indentF 4 (fromString $ show coreCfg)
+      |+ ""
+    Warp.runSettings settings
+      . reqLoggerMiddleware
+      . errLoggerMiddleware
+      . errorJsonWrapMiddleware
+      $ let context = apiKeyAuthHandler (case scServerApiKey serverConfig of Confidential t → apiKeyFromText t) :. EmptyContext
+         in serveWithContext mainAPI context
+              $ hoistServerWithContext
+                mainAPI
+                (Proxy ∷ Proxy '[AuthHandler Wai.Request ()])
+                (\ioAct → Handler . ExceptT $ first (apiErrorToServerError . exceptionHandler) <$> try ioAct)
+              $ mainServer ctx